Get target batch storage config from absolute manifest URLs
